Add a createjail option for setting nameserver
Copying /etc/resolv.conf will not work if the host system runs a name
server (the file is usually empty or at least does not reference any
nameserver in that case).  Add an option to define a nameserver in the
jail before creating the snapshot, and warn the user when creating the
jail in case it is likely not to work as expected.