Create packages on failure.
Put the logic to create packages in a function. By the way, remove this pipe on
awk.
Use this function to package what is installed so far when an error is
reported. This is useful when the build fail while building a dep but we already
built 2 hours worth of deps (true story!)