Remove the queue lock(1)
This was originally done since the pool/ dir had a combination of
'ready to build', 'building' and 'needs deps'. Now that all are
separated, the lockf is not needed. The pkgdirs are atomically
moved between each and atomically cleaned up.

next_in_queue() also only runs from 1 process always.
The cleanup is still in its own clean.sh, but could now be moved
into common.sh.

This change is reported to fix the build getting stuck up to
60 seconds looking for another port to build while others are
cleaning up. If there are X builders and X finish at the same
time, they must serially cleanup, while the master also waits
for the lock to be released. Now the master can pick a new
pkg to build much quicker.