When using TMPFS_ALL, still mount ref/poudriere in tmpfs, which allows cp/pax/future -x to not recopy the dir into the builders
pax(1) claims to support multiple -s, which could be used to ignore /poudriere, but it
does not work right, and is prone to breaking if we stop using pax(1) and forget
to ignore /poudriere. By mounting /poudriere in tmpfs always (only in ref), we ensure
that future copying with -x will not copy it.