MFT: ZFS: Workaround zteardown deadlock on 9.0/9.1-RELEASE when SIGINTing compute_deps
There may be other cases of this. This specific case was parallel_run
compute_deps() executing in the host, reading/writing to the pkgname cache
in the ZFS jail and executing the make -V inside the jail. Somewhere here
was a race condition where it would hold a fd open to a deleted file while
zfs destroy was called.

The fix is to SIGKILL all parallel children before cleanup and wait for
them to exit.