Follow-up to [b65f3c27ce]: Speedup deleting old packages significantly
The problem was that port_is_needed() was called before parallelizing, which
bottlenecked the delete_old_pkg() call. Moving that check into delete_old_pkg()
allows more to run in parallel again and removes the bottleneck.

In a test of bulk -Cf p5.ports, the dependency/delete startup went from
400 seconds to 161 seconds.