Save up to 6 seconds during slave destruction
The previous code blindly tried to umount any possible mount, twice,
with 3 seconds of sleep before attempts.  The new code founds out what
is actually mounted and only attempts to umount those in a single
umount instance.  Additionally, if successful, the 3-second wait is
skipped.  This saves 6 seconds and a lot of known failed umount attempts.