Add "LESS_SANDBOXING" option to poudriere.conf
This new option will tell poudriere to skip using /portdisfiles and
/new_packages directories and just use the collective /distfiles and
/packages directories.

Yes, it takes any some of the sandboxing, but it returns RAM resources
as these areas are TMPFS on DragonFly.  Moreover, the purpose of
portdistfiles is to catch ports that aren't specifying their distfiles.
DragonFly DPorts Builder doesn't have the directive to catch bad ports,
but rather it assumes they are good and tries to build it. 

At this time, the benefit of the "/new_packages" sandbox for DragonFly's
chroot implementation is doubtful; it appears to exist mainly for
FreeBSD's jail implementation that has a Read-Only mount policy.
Therefore, I recommend to use LESS_SANDBOXING=yes on DragonFly.