Major improvements to web interface
These changes fix a lot of problems with the web interface.
The tables now sort without dynamic column resizing.
The font type and sizes are now uniform
The links are no longer underlined.
The categories were put in the same order as the progress bar.
The foreground color of the built/failed/skipped classes were defined.
The background colors of the same were also tweaked.
The H2 class was tweaked.
The useless jump links from header were removed (they caused the totals to
  be unreadable links)
The header summary now has uniform column widths.


The biggest issue, the overlay of the ports tables over the builders table,
has not yet been addressed.  Neither has the dynamic resizing of that table.
It is enough work to deserve a separate commit, so it will come later.