Web interface: Builders table, layout fixes
These changes fix the layout of the web interface.  The Builders table is
now a fixed width (300 px), and the other tables are also fixed at 700 px.
No table will resize; browsers with less than 1050 pixels will display a
horizontal scroll bar.  This permanent fixes the resizing / overlay
problems that required a 1920px monitor to be able to view the data.

The builder table was overhauled.  The "main" status is now a table footer
spanning 3 columns instead of another data row.  The column sorting has
been removed from this table, it only displays in order of increasin ID.
The origin is split across two lines, with the category appearing on the
second line and it's not part of the portsmon hyperlink.
The status column cells also display over two lines where the first line
is the builder status and the second line is reserved for elapsed builder
time (future update).

Javascript, css, and index.html were all modified to make this happen.