Reimplement "Top" link and status links
I recently completely removed the status links because I didn't link how
the totals were turned into links.  So I changed the clickable area from
the text to the table cell.  I also added back the "Top" link in the footer
minus the "top" title which did not have analogues elsewhere.

While it worked exactly as it does it the trunk branch, it did not work
well in my opinion.  The header and footer would often disappear, only to
return if you scroll a bit.  Parts of the screen were missing, maybe they'd
come back in a few seconds, maybe not.  Overall, pretty fragile.

Since we have jquery, I decided to use it.  I changed all the href
relocations to call the new "jump()" function instead.  It works perfectly.
The only thing to note is the section you jump to is not red anymore because
the "target:" css class is not being invoked.

Well, a second note is "Top" changes the url to add "#top" to it, which
I don't really like, so I might tweak this one more time.