Use a shadow directory with hard links during build
This keeps the repository untouched until the very end. It is not quite
atomic yet, but is better than having the repository broken for several
hours or days while a build is running, or has failed.

The repository is only 'committed' if the build does not crash. Failing
packages still will not be in the new repository yet. Keeping old packages
may come later.